--- ui-v2/src/components/automations/automations-wizard/actions-step/action-change-flow-run-state-fields/index.ts
@@ -1,1 +1,0 @@
-export { ActionChangeFlowRunStateFields } from "./action-change-flow-run-state-fields";
