+++ ui-v2/src/components/automations/automations-wizard/actions-step/action-step.tsx
@@ -2,10 +2,10 @@ import { Button } from "@/components/ui/button";
 import { Icon } from "@/components/ui/icons";
 import { Typography } from "@/components/ui/typography";
 import { useWatch } from "react-hook-form";
-import { ActionChangeFlowRunStateFields } from "./action-change-flow-run-state-fields";
 import { ActionsSchema } from "./action-type-schemas";
 import { ActionTypeSelect } from "./action-type-select";
 import { AutomationsSelectStateFields } from "./automations-select-state-fields";
+import { ChangeFlowRunStateFields } from "./change-flow-run-fields";
 
 type ActionStepProps = {
 	index: number;
@@ -45,7 +45,7 @@ const ActionTypeAdditionalFields = ({
 	const actionType = useWatch<ActionsSchema>({ name: `actions.${index}.type` });
 	switch (actionType) {
 		case "change-flow-run-state":
-			return <ActionChangeFlowRunStateFields index={index} />;
+			return <ChangeFlowRunStateFields index={index} />;
 		case "run-deployment":
 		case "pause-deployment":
 		case "resume-deployment":
