+++ ui-v2/src/components/automations/automations-wizard/actions-step/actions-step.test.tsx
@@ -5,6 +5,7 @@ import { Automation } from "@/api/automations";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { buildApiUrl, createWrapper, server } from "@tests/utils";
+import { mockPointerEvents } from "@tests/utils/browser";
 import { http, HttpResponse } from "msw";
 import { beforeAll, describe, expect, it, vi } from "vitest";
 
@@ -17,30 +18,7 @@ describe("ActionsStep", () => {
 		);
 	};
 
-	beforeAll(() => {
-		/**
-		 * JSDOM doesn't implement PointerEvent so we need to mock our own implementation
-		 * Default to mouse left click interaction
-		 * https://github.com/radix-ui/primitives/issues/1822
-		 * https://github.com/jsdom/jsdom/pull/2666
-		 */
-		class MockPointerEvent extends Event {
-			button: number;
-			ctrlKey: boolean;
-			pointerType: string;
-
-			constructor(type: string, props: PointerEventInit) {
-				super(type, props);
-				this.button = props.button || 0;
-				this.ctrlKey = props.ctrlKey || false;
-				this.pointerType = props.pointerType || "mouse";
-			}
-		}
-		window.PointerEvent = MockPointerEvent as never;
-		window.HTMLElement.prototype.scrollIntoView = vi.fn();
-		window.HTMLElement.prototype.releasePointerCapture = vi.fn();
-		window.HTMLElement.prototype.hasPointerCapture = vi.fn();
-	});
+	beforeAll(mockPointerEvents);
 
 	describe("multiple actions", () => {
 		it("able to add multiple actions", async () => {
