+++ ui-v2/src/components/automations/automations-wizard/automations-trigger-template-select/automations-trigger-template-select.test.tsx
@@ -1,33 +1,12 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import { mockPointerEvents } from "@tests/utils/browser";
 import { expect, test, vi } from "vitest";
 
 import { AutomationsTriggerTemplateSelect } from "./automations-trigger-template-select";
 
 test("AutomationsTriggerTemplateSelect can select an option", async () => {
-	/**
-	 * JSDOM doesn't implement PointerEvent so we need to mock our own implementation
-	 * Default to mouse left click interaction
-	 * https://github.com/radix-ui/primitives/issues/1822
-	 * https://github.com/jsdom/jsdom/pull/2666
-	 */
-	class MockPointerEvent extends Event {
-		button: number;
-		ctrlKey: boolean;
-		pointerType: string;
-
-		constructor(type: string, props: PointerEventInit) {
-			super(type, props);
-			this.button = props.button || 0;
-			this.ctrlKey = props.ctrlKey || false;
-			this.pointerType = props.pointerType || "mouse";
-		}
-	}
-	window.PointerEvent = MockPointerEvent as never;
-	window.HTMLElement.prototype.scrollIntoView = vi.fn();
-	window.HTMLElement.prototype.releasePointerCapture = vi.fn();
-	window.HTMLElement.prototype.hasPointerCapture = vi.fn();
-
+	mockPointerEvents();
 	const user = userEvent.setup();
 
 	// ------------ Setup
